+++ CHANGELOG.md
@@ -5,12 +5,29 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 ### Added
+
+### Changed
+
+### Fixed
+- Centralized the instantiation of the database engine and session (`clients/instances.py`) to improve geo-sensemaker performance. 
+- Pulled out unnecessary processing from thread locks in Geo controller, while maintaining safety of buffers. 
+
+- Sync client unable to connect due to CA File ssl auth errors
+
+### Removed
+
+## [0.15.0] - 2025-10-21
+
+### Added
+- Async OMS Client
 - Ensure automated tests have 80% coverage
 - Added Pydantic models and a TypeAdapter to validate the `audit_log_error_acm` loaded from JSON.
 - OpenTelemetry integration for queue processing time and event counting metrics
+- Filter Mil Symbol Attribute Inputs based on IRI before we rehydrate them
 
 ### Changed
 - Removed Postgres Client and EPEL repository installation in Dockerfile
+- Updated ATOMS dependency to Grimlock-INC-33
 
 ### Fixed
 
@@ -18,6 +35,7 @@ All notable changes to this project will be documented in this file.
 - Corrected the Jenkins build job for the Sensemaking image which occasionaly created incorrect versions
 - Allow test coverage portion of the Jenkins pipeline to run and post results to SonarQube in AIDE
 - Fixed MilSymbol Sensemaker bug where an invalid character part of a symbol Id code would be processed and cause errors in the service
+- Fixed issue with the open telemetry unit tests not work propererly when the services weren't running
 
 ### Removed